requirements.txt
aerich==0.7.2
aioredis==2.0.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
beautifulsoup4==4.12.2
certifi==2023.7.22
click==8.1.7
databases==0.8.0
dictdiffer==0.9.0
Faker==19.13.0
fastapi==0.104.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pypika-tortoise==0.1.6
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.49
starlette==0.27.0
text-unidecode==1.3
tomlkit==0.12.1
tortoise-orm==0.20.0
typing_extensions==4.8.0
uvicorn==0.23.2
xmltodict==0.13.0
boto3==1.35.45
